pyproject.toml
[project]
name = "research_crew"
version = "0.1.0"
description = "research_crew using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.159.0,<1.0.0",
    "arxiv>=2.2.0",
    "crewai>=0.159.0",
    "crewai-tools[mcp]>=0.62.0",
    "ipykernel>=6.30.1",
    "langchain-google-genai>=2.1.9",
    "langchain-groq>=0.3.7",
    "mcp>=1.12.4",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
    "streamlit>=1.48.1",
]

[project.scripts]
research_crew = "research_crew.main:run"
run_crew = "research_crew.main:run"
train = "research_crew.main:train"
replay = "research_crew.main:replay"
test = "research_crew.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
